requirements.txt
mysqlclient==2.2.7

argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0

djangorestframework==3.16.0
djangorestframework-api-response==0.1.0
djangorestframework-simplejwt==5.3.0

boto3==1.39.12
botocore==1.39.12
eth-account==0.13.7
channels==4.1.0
confluent-kafka==2.11.0
Django==5.2
django-ckeditor==6.7.3
django-axes==8.0.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-redis==6.0.0
django-storages==1.14.6
jsonpickle==4.1.1
djoser==2.3.3
web3==7.12.1
Pillow==11.3.0
psycopg2-binary==2.9.10
python-slugify==8.0.1
PyJWT==2.10.1
stripe==12.3.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0

gunicorn==23.0.0
whitenoise==6.9.0
uvicorn==0.35.0

django-unslashed==0.3.0
openpyxl==3.1.5
python-dotenv==1.1.1